# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ChattyCathy API
  description: |
    ChattyCathy API Server with JWT authentication and RBAC.

    ## Authentication

    This API uses JWT (RS256) authentication with access and refresh tokens.

    - **Access tokens**: Short-lived (15 min), used in Authorization header
    - **Refresh tokens**: Long-lived (7 days), used to obtain new access tokens

    ### Authentication Methods

    - **Google OAuth**: Sign in with Google (recommended for users)
    - **Username/Password**: Traditional login (for demo/testing)

    ### For Web Clients
    Tokens are automatically set as HTTP-only cookies after login.

    ### For Mobile/API Clients
    Use the `Authorization: Bearer <token>` header with access tokens.

    ## Role-Based Access Control (RBAC)

    The API implements permission-based access control:

    ### Default Roles
    - **admin**: Full access to all resources
    - **user**: Basic read access (ping, news) - *automatically assigned to new users*
    - **editor**: Read and write access to content

    ### Default Permissions
    - `ping:read` - Access ping endpoint
    - `news:read`, `news:create`, `news:update`, `news:delete`
    - `users:read`, `users:create`, `users:update`, `users:delete`
    - `roles:read`, `roles:create`, `roles:update`, `roles:delete`

    ### Automatic Role Assignment
    New users registered via Google OAuth are automatically assigned the **user** role,
    which grants `ping:read` and `news:read` permissions.

...
servers:
...
paths:
...
  /auth/google:
    post:
      summary: Authenticate with Google
      description: |
        Authenticate using a Google OAuth access token or authorization code.

        **Implicit Flow (recommended for SPAs):**
        Send the `access_token` received from Google's OAuth popup.

        **Authorization Code Flow:**
        Send the `code` received from Google's redirect.

        On success, returns JWT tokens and user info. A new user is created if they don't exist.
        Users are assigned the "user" role by default.
      operationId: googleAuth
      tags:
        - google
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleAuthRequest"
      responses:
        "200":
          description: Authentication successful
          headers:
            Set-Cookie:
              description: HTTP-only cookies for access_token and refresh_token
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request - access_token or code required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid Google credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to process user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/refresh:
    post:
      summary: Refresh tokens
      description: |
        Get new access and refresh tokens using a valid refresh token.

        The refresh token can be provided via:
        - HTTP-only cookie (web clients)
        - Request body (`refresh_token` field)
        - Header (`X-Refresh-Token`)

        **Note**: Refresh tokens are rotated on each use. The old refresh token becomes invalid.
      operationId: refresh
      tags:
        - auth
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens refreshed successfully
          headers:
            Set-Cookie:
              description: New HTTP-only cookies for access_token and refresh_token
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /protected/secret:
    get:
      summary: Secret endpoint
      description: A protected endpoint that requires authentication
      operationId: getSecret
      tags:
        - protected
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Secret data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /protected/profile:
    get:
      summary: Get user profile
      description: Get the current user's profile information
      operationId: getProfile
      tags:
        - protected
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    GoogleAuthRequest:
      type: object
      properties:
        access_token:
          type: string
          description: Google OAuth access token (for implicit flow)
        code:
          type: string
          description: Google OAuth authorization code (for auth code flow)
      description: Either access_token or code must be provided
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
        access_token_expires_in:
          type: integer
          description: Access token expiry in seconds
          example: 900
        token_type:
          type: string
          example: Bearer
        user:
          $ref: "#/components/schemas/User"
      required:
        - access_token
        - refresh_token
        - access_token_expires_in
        - token_type
        - user
    User:
      type: object
      properties:
        id:
          type: integer
          description: User ID
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: John Doe
        picture:
          type: string
          format: uri
          description: Profile picture URL
          example: "https://lh3.googleusercontent.com/..."
        role:
          type: string
          description: User's role
          example: user
        permissions:
          type: array
          items:
            type: string
          description: List of user's permissions
          example: ["ping:read", "news:read"]
      required:
        - id
        - email
        - name
        - picture
        - role
...
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: The refresh token (if not using cookies)
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Access token expiry in seconds
          example: 900
        refresh_expires_in:
          type: integer
          description: Refresh token expiry in seconds
          example: 604800
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
        - refresh_expires_in
...
    SecretResponse:
      type: object
      properties:
        message:
          type: string
          example: This is a protected secret!
        user:
          type: string
          description: User ID from the JWT
        role:
          type: string
          description: User role from the JWT
      required:
        - message
        - user
        - role
    ProfileResponse:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
      required:
        - user_id
        - email
        - role
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...